KFP Client split into client and pipeline generator
