+++ kedro_kubeflow/kfpclient.py
@@ -1,21 +1,15 @@
 import json
 import logging
-import os
-import re
 import uuid
-from functools import wraps
-from inspect import Parameter, signature
 from tempfile import NamedTemporaryFile
-from typing import Dict, Set
 
-from kedro.pipeline.node import Node
-from kfp import Client, dsl
+from kfp import Client
 from kfp.compiler import Compiler
-from kubernetes.client import V1EnvVar, V1SecurityContext
 from tabulate import tabulate
 
-from .auth import IAP_CLIENT_ID, AuthHandler
-from .utils import is_mlflow_enabled
+from .auth import AuthHandler
+from .generator import PipelineGenerator
+from .utils import clean_name
 
 WAIT_TIMEOUT = 24 * 60 * 60
 
@@ -29,9 +23,7 @@ class KubeflowClient(object):
         self.host = config.host
         self.client = Client(self.host, existing_token=token)
         self.project_name = project_name
-        self.context = context
-        dsl.ContainerOp._DISABLE_REUSABLE_COMPONENT_WARNING = True
-        self.volume_meta = config.run_config.volume
+        self.generator = PipelineGenerator(config, project_name, context)
 
     def list_pipelines(self):
         pipelines = self.client.list_pipelines(page_size=30).pipelines
@@ -49,7 +41,9 @@ class KubeflowClient(object):
         image_pull_policy="IfNotPresent",
     ) -> None:
         run = self.client.create_run_from_pipeline_func(
-            self.generate_pipeline(pipeline, image, image_pull_policy),
+            self.generator.generate_pipeline(
+                pipeline, image, image_pull_policy
+            ),
             arguments={},
             experiment_name=experiment_name,
             run_name=run_name,
@@ -58,155 +52,21 @@ class KubeflowClient(object):
         if wait:
             run.wait_for_run_completion(timeout=WAIT_TIMEOUT)
 
-    def generate_pipeline(self, pipeline, image, image_pull_policy):
-        def _customize_op(op):
-            op.container.set_image_pull_policy(image_pull_policy)
-            if self.volume_meta and self.volume_meta.owner is not None:
-                op.container.set_security_context(
-                    V1SecurityContext(run_as_user=self.volume_meta.owner)
-                )
-            return op
-
-        def maybe_add_params(kedro_parameters):
-            def decorator(f):
-                @wraps(f)
-                def wrapper(*args, **kwargs):
-                    return f()
-
-                sig = signature(f)
-                new_params = (
-                    Parameter(name, Parameter.KEYWORD_ONLY, default=default)
-                    for name, default in kedro_parameters.items()
-                )
-                wrapper.__signature__ = sig.replace(parameters=new_params)
-                return wrapper
-
-            return decorator
-
-        @dsl.pipeline(
-            name=self.project_name,
-            description="Kubeflow pipeline for Kedro project",
-        )
-        @maybe_add_params(self.context.params)
-        def convert_kedro_pipeline_to_kfp() -> None:
-            """Convert from a Kedro pipeline into a kfp container graph."""
-
-            node_volumes = (
-                _setup_volumes() if self.volume_meta is not None else {}
-            )
-            node_dependencies = self.context.pipelines.get(
-                pipeline
-            ).node_dependencies
-            kfp_ops = _build_kfp_ops(node_dependencies, node_volumes)
-            for node, dependencies in node_dependencies.items():
-                for dependency in dependencies:
-                    kfp_ops[node.name].after(kfp_ops[dependency.name])
-
-        def _setup_volumes():
-            vop = dsl.VolumeOp(
-                name="data-volume-create",
-                resource_name="data-volume",
-                size=self.volume_meta.size,
-                modes=self.volume_meta.access_modes,
-                storage_class=self.volume_meta.storageclass,
-            )
-            if self.volume_meta.skip_init:
-                return {"/home/kedro/data": vop.volume}
-            else:
-                volume_init = _customize_op(
-                    dsl.ContainerOp(
-                        name="data-volume-init",
-                        image=image,
-                        command=["sh", "-c"],
-                        arguments=[
-                            " ".join(
-                                [
-                                    "cp",
-                                    "--verbose",
-                                    "-r",
-                                    "/home/kedro/data/*",
-                                    "/home/kedro/datavolume",
-                                ]
-                            )
-                        ],
-                        pvolumes={"/home/kedro/datavolume": vop.volume},
-                    )
-                )
-                return {"/home/kedro/data": volume_init.pvolume}
-
-        def _build_kfp_ops(
-            node_dependencies: Dict[Node, Set[Node]], node_volumes: Dict
-        ) -> Dict[str, dsl.ContainerOp]:
-            """Build kfp container graph from Kedro node dependencies. """
-            kfp_ops = {}
-
-            iap_env_var = V1EnvVar(
-                name=IAP_CLIENT_ID, value=os.environ.get(IAP_CLIENT_ID, "")
-            )
-            nodes_env = [iap_env_var]
-
-            if is_mlflow_enabled():
-                kfp_ops["mlflow-start-run"] = _customize_op(
-                    dsl.ContainerOp(
-                        name="mlflow-start-run",
-                        image=image,
-                        command=["kedro"],
-                        arguments=[
-                            "kubeflow",
-                            "mlflow-start",
-                            dsl.RUN_ID_PLACEHOLDER,
-                        ],
-                        container_kwargs={"env": [iap_env_var]},
-                        file_outputs={"mlflow_run_id": "/tmp/mlflow_run_id"},
-                    )
-                )
-
-                nodes_env.append(
-                    V1EnvVar(
-                        name="MLFLOW_RUN_ID",
-                        value=kfp_ops["mlflow-start-run"].output,
-                    )
-                )
-
-            for node in node_dependencies:
-                name = _clean_name(node.name)
-                params = ",".join(
-                    [
-                        f"{param}:{dsl.PipelineParam(param)}"
-                        for param in self.context.params.keys()
-                    ]
-                )
-                kfp_ops[node.name] = _customize_op(
-                    dsl.ContainerOp(
-                        name=name,
-                        image=image,
-                        command=["kedro"],
-                        arguments=[
-                            "run",
-                            "--params",
-                            params,
-                            "--node",
-                            node.name,
-                        ],
-                        pvolumes=node_volumes,
-                        container_kwargs={"env": nodes_env},
-                    )
-                )
-
-            return kfp_ops
-
-        return convert_kedro_pipeline_to_kfp
-
     def compile(
         self, pipeline, image, output, image_pull_policy="IfNotPresent"
     ):
         Compiler().compile(
-            self.generate_pipeline(pipeline, image, image_pull_policy), output
+            self.generator.generate_pipeline(
+                pipeline, image, image_pull_policy
+            ),
+            output,
         )
         self.log.info("Generated pipeline definition was saved to %s" % output)
 
     def upload(self, pipeline, image, image_pull_policy="IfNotPresent"):
-        pipeline = self.generate_pipeline(pipeline, image, image_pull_policy)
+        pipeline = self.generator.generate_pipeline(
+            pipeline, image, image_pull_policy
+        )
 
         if self._pipeline_exists(self.project_name):
             pipeline_id = self._get_pipeline_id(self.project_name)
@@ -250,7 +110,7 @@ class KubeflowClient(object):
     def _upload_pipeline_version(
         self, pipeline_func, pipeline_id, pipeline_name
     ):
-        version_name = f"{_clean_name(pipeline_name)}-{uuid.uuid4()}"[:100]
+        version_name = f"{clean_name(pipeline_name)}-{uuid.uuid4()}"[:100]
         with NamedTemporaryFile(suffix=".yaml") as f:
             Compiler().compile(pipeline_func, f.name)
             return self.client.pipeline_uploads.upload_pipeline_version(
@@ -306,7 +166,3 @@ class KubeflowClient(object):
             for job in my_runs:
                 self.client.jobs.delete_job(job.id)
                 self.log.info(f"Previous schedule deleted {job.id}")
-
-
-def _clean_name(name: str) -> str:
-    return re.sub(r"[\W_]+", "-", name).strip("-")
